requirements.txt
fastapi
uvicorn[standard]
pymongo
redis
aiokafka
pyyaml